+++ sscRaft/geometries/methods.py
@@ -34,10 +34,11 @@ def fbp(tomogram, angles = None, obj = None, dic = None, **kwargs):
         * ``dic['filter']`` (str,optional): Filter type [Default: \'lorentz\']
 
             #. Options = (\'none\',\'gaussian\',\'lorentz\',\'cosine\',\'rectangle\',\'hann\',\'hamming\',\'ramp\')
-          
+        
+        * ``dic['detectorPixel[m]']`` (float,optional): Detector pixel size in meters [Default: 1.0]
         * ``dic['beta/delta']`` (float,optional): Paganin by slices method ``beta/delta`` ratio [Default: 0.0 (no Paganin applied)]
-        * ``dic['z2[m]']`` (float,optional): Sample-Detector distance in meters used on Paganin by slices method. [Default: 0.0 (no Paganin applied)]
-        * ``dic['energy[eV]']`` (float,optional): beam energy in eV used on Paganin by slices method. [Default: 0.0 (no Paganin applied)]
+        * ``dic['z2[m]']`` (float,optional): Sample-Detector distance in meters used on Paganin by slices method. [Default: 1.0]
+        * ``dic['energy[eV]']`` (float,optional): beam energy in eV used on Paganin by slices method. [Default: 1.0 ]
         * ``dic['regularization']`` (float,optional): Regularization value for filter ( value >= 0 ) [Default: 1.0]
         * ``dic['padding']`` (int,optional): Data padding - Integer multiple of the data size (0,1,2, etc...) [Default: 2]
         * ``dic['blocksize']`` (int,optional): Block of slices to be simulteneously computed [Default: 0 (automatically)]
@@ -50,12 +51,11 @@ def fbp(tomogram, angles = None, obj = None, dic = None, **kwargs):
     
     """
     required = ('gpu',)        
-    optional = ( 'filter','rotation axis offset','padding','regularization','beta/delta','blocksize','energy[eV]','z2[m]','method')
-    default  = ('lorentz',                     0,        2,             1.0,         0.0,          0,         1.0,    1.0,    'RT')
+    optional = ( 'filter','rotation axis offset','padding','regularization','beta/delta','blocksize','energy[eV]','z2[m]','method', 'detectorPixel[m]')
+    default  = ('lorentz',                     0,        2,             1.0,         0.0,          0,         1.0,    1.0,    'RT',                1.0)
     
     dic = SetDictionary(dic,required,optional,default)
 
-    # dic['regularization'] = 1.0
     method                = dic['method']
     gpus                  = dic['gpu']