+++ sscRaft/processing/alignments/rotationaxis.py
@@ -215,10 +215,38 @@ def correct_rotation_axis(data: numpy.ndarray, deviation: int) -> numpy.ndarray:
     else:
         proj[:,:,0:data.shape[2]] = data
 
-    # logger.info(f'Corrected projection for rotation axis: new shape {proj.shape}')
-    logger.info(f'Corrected projection for rotation axis: modified')
+    logger.info(f'Corrected projection for rotation axis: new shape {proj.shape}')
 
-    
-    proj = proj[:,:,numpy.abs(deviation):-numpy.abs(deviation)]
+    return proj
+
+def correct_rotation_axis_cropped(data: numpy.ndarray, deviation: int) -> numpy.ndarray:
+    """Corrects the rotation axis of a data according to a deviation value defined 
+    by the number of pixels translated form the center of the data.
+
+    The deviation value, with sign, is computed through the sscRaft function ``Centersino()``.
+
+    Args:
+        data (ndarray): Projection tomogram. The axes are [slices, angles, lenght]
+        deviation (int): Number of pixels representing the rotation axis deviation
+
+    Returns:
+        (ndarray): Rotation axis corrected tomogram (3D) with shape [slices, angles, lenght] 
+
+    * CPU function
+    """
+    logger.info(f'Applying given rotation axis correction deviation value: {deviation}')
+
+    deviation = - deviation # Fix centersino value
+
+    proj = numpy.zeros((data.shape[0], data.shape[1], data.shape[2] + 2 * numpy.abs(deviation)))
+
+    if(deviation < 0):
+        proj[:,:,2 * numpy.abs(deviation):data.shape[2] + 2 * numpy.abs(deviation)] = data
+    else:
+        proj[:,:,0:data.shape[2]] = data
+
+
+    if deviation != 0:
+        proj = proj[:,:,numpy.abs(deviation):-numpy.abs(deviation)]
 
     return proj